refactored in the structure, started adding authentication stuff
