+++ lib/omsmodule.js
@@ -5,9 +5,24 @@
 
 var express = require('express');
 var app     = express();
-var port    =   process.env.PORT || 8081;
+var port    = process.env.PORT || 8081;
 
 var exphbs = require('express-handlebars');
+
+//var passport = require('passport');
+//var flash    = require('connect-flash');
+
+//var morgan       = require('morgan');
+//var cookieParser = require('cookie-parser');
+//app.use(morgan('dev')); // log every request to the console
+//app.use(cookieParser()); // read cookies (needed for auth)
+
+//var session      = require('express-session');
+//app.use( session( { secret: 'ilovescotchscotchyscotchscotch' } ) ); // session secret
+//app.use( passport.initialize() );
+//app.use( passport.session() ); // persistent login sessions
+//app.use( flash() ); // use connect-flash for flash messages stored in session
+
 
 var bodyParser = require('body-parser');
 app.use(bodyParser.json()); // support json encoded bodies
@@ -21,8 +36,8 @@ var restClient = restify.createJsonClient({
 });
 
 
-var routes = require('./routes.js');
-
+var routes = require('./app/routes.js');
+//require('./config/passport')(passport)
 
 // VIEWS ENGINE
 // ==============================================
@@ -71,6 +86,8 @@ router.get('/user/:uid', routes.getMember, routes.getMemberships, routes.renderP
 
 
 router.get('/login', routes.renderLogin );
+
+//router.post('/login', routes.doLogin );
 
 
 router.get('/applicants/:bodycode', routes.getApplicationsToBody, routes.renderApplicationsToBody );
