+++ lib/public/fonts/glyphicons-halflings-regular.woff2
