+++ lib/public/js/npm.js
