+++ lib/views/layouts/single.hbs
