+++ lib/views/partials/footer.hbs
