+++ lib/views/partials/head.hbs
