+++ lib/views/partials/navbar.hbs
