+++ lib/views/register.hbs
