+++ lib/views/registration_success.hbs
